# Fixes for tp_new for C extensions