Split from https://github.com/dropbox/pyston/pull/1014

In the second commit, I add a special case in `typeCallInner` for `int_cls`, since it's the most common one to show up. It helps performance a bit.
```
# without special case
Comparing to '['baseline']'
              pyston (calibration)                      :   1.01s baseline: 1.02 ( -0.6%)
              pyston django_template3.py                :   1.98s baseline: 1.97 (  0.2%)
              pyston pyxl_bench.py                      :   1.63s baseline: 1.62 (  0.5%)
              pyston sqlalchemy_imperative2.py          :   2.12s baseline: 2.11 (  0.3%)
              pyston django_template3_10x.py            :  11.72s baseline: 11.53 (  1.6%)
              pyston pyxl_bench_10x.py                  :  12.84s baseline: 12.78 (  0.5%)
              pyston sqlalchemy_imperative2_10x.py      :  16.31s baseline: 16.13 (  1.1%)
              pyston (geomean-2ec9)                     :   1.90s baseline: 1.89 (  0.3%)
# with special case
Comparing to '['baseline']'
              pyston (calibration)                      :   1.02s baseline: 1.02 (  0.6%)
              pyston django_template3.py                :   1.99s baseline: 1.97 (  0.7%)
              pyston pyxl_bench.py                      :   1.63s baseline: 1.62 (  0.7%)
              pyston sqlalchemy_imperative2.py          :   2.11s baseline: 2.11 ( -0.1%)
              pyston django_template3_10x.py            :  11.52s baseline: 11.53 ( -0.0%)
              pyston pyxl_bench_10x.py                  :  12.78s baseline: 12.78 (  0.0%)
              pyston sqlalchemy_imperative2_10x.py      :  16.15s baseline: 16.13 (  0.1%)
              pyston (geomean-2ec9)                     :   1.90s baseline: 1.89 (  0.4%)
```

But I don't know how I feel about it, it seems quite hacky for a relatively small performance drop.