@kmod 2016-01-25
I think if we do this right, there shouldn't be any perf change, since in theory we should be able to get it to the point that these new functions don't get called at all in the benchmarks.

One way to do this would be to have a flag that says "hey even though I have tp_new set, I also have a Python `__new__` available so you should try calling that".  We already have two classes that want this flag (all the exception types, and int), but there are potentially more.  I think the 'is_pyston_class' flag is actually already suitable for this, though maybe a bit hacky.

Another approach is to switch this to the "rewriter aggressiveness backoff" system which is designed for this sort of thing: when we have multiple options for how to try to execute things, but we don't know how many of the fastpaths we'll be able to take.  I think this makes sense, but it's probably overkill and we can keep the change much smaller by checking is_pyston_class.

If you have time, it'd be nice to test putting a stat counter inside that special case, and verify that it gets triggered the same number of times before and after the PR.