@rudi-c 2016-02-13
I added the noexcept clause.

The performance drop that I initially observed seem to be gone now and the special case in `typeCallInner` hack no longer does anything, so I removed it.

```
Comparing to '['baseline']'
              pyston (calibration)                      :   1.02s baseline: 1.02 ( -0.4%)
              pyston django_template3.py                :   2.08s baseline: 2.08 ( -0.1%)
              pyston pyxl_bench.py                      :   1.80s baseline: 1.82 ( -1.1%)
              pyston sqlalchemy_imperative2.py          :   2.16s baseline: 2.16 ( -0.1%)
              pyston django_template3_10x.py            :  12.33s baseline: 12.27 (  0.5%)
              pyston pyxl_bench_10x.py                  :  13.87s baseline: 13.82 (  0.4%)
              pyston sqlalchemy_imperative2_10x.py      :  16.60s baseline: 16.67 ( -0.4%)
              pyston (geomean-2ec9)                     :   2.01s baseline: 2.02 ( -0.4%)
```

The new case does get called in the benchmarks, setting `is_pyston_class` doesn't change that, but it doesn't improve performance either (it actually makes it worse). But that does lead me to wonder, was there any reason not to set `is_pyston_class` to all of our classes? Performance?