@undingen 2016-02-25
Hi, sorry for the very late review!
I added a minor comment but decided I don't want to introduce more overhead and even more delays to the whole process and merged it in and created #1115 to fix it. @Daetalus let me know that this fixes some numpy crashes for example ```test_abc.py``` :-) 
Thanks for the patch!