@undingen 2016-02-25
Oh I forgot to mention: I verified that this does not decrease the performance even though this will cause more tp_new calls...